// src/app/feature-modules/main/sidebarconfigure/file-routing/data/constant-value.enum.ts
export enum Constant {
    CONTROLLER = "/SymFileRouting",
    PATH = "/main/configure/file-routing",
    UNCLICKABLE = "unclickable",
    CLICKABLE = "clickable",
    CREATE = "/create",
    EDIT = "/edit",
    DELETE = "/delete",
    SORT_BY = "trigger_id",
    SORT_ORDER = "asc",
    FILTER_BY = "trigger_id",
    SAVE = "Save",
    UPDATE = "Update"
}